# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  enums.InstallmentStatus:
    enum:
    - PENDING
    - PAID
    - OVERDUE
    type: string
    x-enum-varnames:
    - Pending
    - Paid
    - Overdue
...
  request.CreateInstallmentRequest:
    properties:
      amount:
        type: number
      credit_account_id:
        type: integer
      due_date:
        type: string
      status:
        $ref: '#/definitions/enums.InstallmentStatus'
    required:
    - amount
    - credit_account_id
    - due_date
    - status
    type: object
  request.CreateLateFeeRequest:
    properties:
      amount:
        type: number
      credit_account_id:
        type: integer
    required:
    - amount
    - credit_account_id
    type: object
...
  request.UpdateInstallmentRequest:
    properties:
      amount:
        type: number
      due_date:
        type: string
      status:
        $ref: '#/definitions/enums.InstallmentStatus'
    type: object
  request.UpdateLateFeeRequest:
    properties:
      amount:
        type: number
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  response.InstallmentResponse:
    properties:
      amount:
        type: number
      credit_account_id:
        type: integer
      due_date:
        type: string
      id:
        type: integer
      status:
        $ref: '#/definitions/enums.InstallmentStatus'
    type: object
  response.LateFeeResponse:
    properties:
      amount:
        type: number
      applied_date:
        type: string
      credit_account_id:
        type: integer
      id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for managing finances in small businesses.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Final Assignment Finance API Rest
  version: "1.0"
paths:
...
  /api/v1/credit-accounts/{creditAccountID}/installments:
    get:
      description: Retrieves all installments for a specific credit account.
      parameters:
      - description: Credit Account ID
        in: path
        name: creditAccountID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.InstallmentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get installments by credit account ID
      tags:
      - Installments
  /api/v1/credit-accounts/{creditAccountID}/installments/overdue:
    get:
      description: Retrieves all overdue installments for a specific credit account.
      parameters:
      - description: Credit Account ID
        in: path
        name: creditAccountID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.InstallmentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get overdue installments by credit account ID
      tags:
      - Installments
  /api/v1/credit-accounts/{creditAccountID}/late-fees:
    get:
      description: Retrieves all late fees for a specific credit account.
      parameters:
      - description: Credit Account ID
        in: path
        name: creditAccountID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.LateFeeResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get late fees by credit account ID
      tags:
      - LateFees
...
  /api/v1/installments:
    post:
      consumes:
      - application/json
      description: Creates a new installment for a credit account.
      parameters:
      - description: Installment details
        in: body
        name: installment
        required: true
        schema:
          $ref: '#/definitions/request.CreateInstallmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.InstallmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a new installment
      tags:
      - Installments
  /api/v1/installments/{id}:
    delete:
      description: Deletes an installment by its ID.
      parameters:
      - description: Installment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete an installment
      tags:
      - Installments
    get:
      description: Retrieves an installment by its ID.
      parameters:
      - description: Installment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.InstallmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get installment by ID
      tags:
      - Installments
    put:
      consumes:
      - application/json
      description: Updates an existing installment by its ID.
      parameters:
      - description: Installment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated installment details
        in: body
        name: installment
        required: true
        schema:
          $ref: '#/definitions/request.UpdateInstallmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.InstallmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update an installment
      tags:
      - Installments
...
  /api/v1/late-fees:
    post:
      consumes:
      - application/json
      description: Creates a new late fee for a credit account.
      parameters:
      - description: Late fee details
        in: body
        name: lateFee
        required: true
        schema:
          $ref: '#/definitions/request.CreateLateFeeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.LateFeeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a late fee
      tags:
      - LateFees
  /api/v1/late-fees/{id}:
    delete:
      description: Deletes a late fee by its ID.
      parameters:
      - description: Late Fee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete a late fee
      tags:
      - LateFees
    get:
      description: Retrieves a late fee by its ID.
      parameters:
      - description: Late Fee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.LateFeeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get late fee by ID
      tags:
      - LateFees
    put:
      consumes:
      - application/json
      description: Updates an existing late fee by its ID.
      parameters:
      - description: Late Fee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated late fee details
        in: body
        name: lateFee
        required: true
        schema:
          $ref: '#/definitions/request.UpdateLateFeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.LateFeeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update a late fee
      tags:
      - LateFees
...
swagger: "2.0"